% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{mcmc_across_parameters}
\alias{mcmc_across_parameters}
\title{Run MCMC with different parameters (possibly with multiple reps fpr each experiment).}
\usage{
mcmc_across_parameters(
  model,
  model_parameter_lists,
  truth = NULL,
  number_of_reps = 100,
  initial_values = list(),
  future_plan = "sequential",
  results_directory = getwd(),
  algorithm_parameter_list = list(),
  external_packages = c(),
  julia_bin_dir = "",
  julia_required_libraries = c(),
  verify_cpp_function_types = FALSE,
  keep_temporary_model_code = FALSE,
  seed = NULL,
  calculate_statistics = TRUE
)
}
\arguments{
\item{model}{A file containing the model.}

\item{model_parameter_lists}{A list of lists containing parameters for the model.}

\item{truth}{(optional) A data frame containing the true values of the parameters (default is NULL, in which case no bias or mean square errors will be calculated).}

\item{number_of_reps}{(optional) The number of chains (default = 100)}

\item{initial_values}{(optional) A list of lists containing the initial values for each rep.}

\item{future_plan}{(optional) The type of parallelism over reps for the future package (default is sequential).}

\item{results_directory}{(optional) The base name of the directories to which results will be written.}

\item{algorithm_parameter_list}{(optional) A list containing named parameters for the algorithm.}

\item{external_packages}{(optional) A vector of names of other R packages the functions rely on.}

\item{julia_bin_dir}{(optional) The directory containing the Julia bin file - only needed if Julia functions are used.}

\item{julia_required_libraries}{(optional) Vector of strings, each of which is a Julia packge that will be installed and loaded.}

\item{verify_cpp_function_types}{(optional) If TRUE, check the types of the parameters of user-defined .cpp functions. If FALSE (default), types are not checked.}

\item{keep_temporary_model_code}{(optional) If FALSE (default), the .cpp file generated for compilation is deleted. If TRUE, this file is left in the working directory.}

\item{seed}{(optional) A seed for the random number generator.}

\item{calculate_statistics}{(optional) Calculate MCMC statistics for each set of parameters.}
}
\value{
Nothing: output can be found in the output_directory.
}
\description{
Run MCMC with different parameters (possibly with multiple reps fpr each experiment).
}
