% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{mcmc_marginal_statistics}
\alias{mcmc_marginal_statistics}
\title{Calculate marginal statistics from MCMC run.}
\usage{
mcmc_marginal_statistics(output, chain = 1, rep = 1, external_index = 1)
}
\arguments{
\item{output}{MCMC output.}

\item{chain}{(optional) The chain for which to calculate the statistics.}

\item{rep}{(optional) The rep (if present) for which to calculate the statistics.}

\item{external_index}{(optional) The external_index (if present) for which to calculate the statistics.}
}
\value{
A data frame containing marginal statistics of the run.
}
\description{
Calculate marginal statistics from MCMC run.
}
