% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{mcmc_statistics}
\alias{mcmc_statistics}
\title{Calculate statistics from an MCMC run.}
\usage{
mcmc_statistics(output, chain = 1, rep = 1, external_index = 1)
}
\arguments{
\item{output}{MCMC output.}

\item{chain}{(optional) The chain for which to calculate the multiESS.}

\item{rep}{(optional) The rep (if present) for which to calculate the multiESS.}

\item{external_index}{(optional) The external_index (if present) for which to calculate the multiESS.}
}
\value{
A data frame containing statistics of the run.
}
\description{
Calculate statistics from an MCMC run.
}
