% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{mcmc_statistics_across_models}
\alias{mcmc_statistics_across_models}
\title{Calculating statistics from MCMC runs over a range of different models (with the same parameters for each model).}
\usage{
mcmc_statistics_across_models(
  models,
  model_parameter_lists,
  results_directory = getwd(),
  truth = NULL
)
}
\arguments{
\item{models}{A list containing the different model files.}

\item{model_parameter_lists}{A list of lists containing parameters for each model.}

\item{results_directory}{(optional) The base name of the directories to which results will be written.}

\item{truth}{(optional) A data frame containing the true values of the parameters (default is NULL, in which case no bias or mean square errors will be calculated).}
}
\value{
Nothing: output can be found in the output_directory.
}
\description{
Calculating statistics from MCMC runs over a range of different models (with the same parameters for each model).
}
