% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{mcmc_statistics_over_reps}
\alias{mcmc_statistics_over_reps}
\title{Load reps of MCMC output.}
\usage{
mcmc_statistics_over_reps(results_directory = getwd(), truth = NULL)
}
\arguments{
\item{results_directory}{(optional) The base name of the directories to which results have been written.}

\item{truth}{(optional) A data frame containing the true values of the parameters (default is NULL, in which case no bias or mean square errors will be calculated).}
}
\value{
A data frame containing the across-rep statistics.
}
\description{
Load reps of MCMC output.
}
