% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{monte_carlo_bias}
\alias{monte_carlo_bias}
\title{Find the bias across Monte Carlo runs.}
\usage{
monte_carlo_bias(per_chain_output, truth)
}
\arguments{
\item{per_chain_output}{Estimate calculated from each chain MCMC in output.}

\item{truth}{The true value being estimated.}
}
\value{
A vector giving the bias for each estimate across the chains.
}
\description{
Find the bias across Monte Carlo runs.
}
