% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{monte_carlo_mean}
\alias{monte_carlo_mean}
\title{Find the mean across Monte Carlo runs.}
\usage{
monte_carlo_mean(per_chain_output)
}
\arguments{
\item{per_chain_output}{Estimate calculated from each chain MCMC in output.}
}
\value{
A vector giving the mean for each estimate across the chains.
}
\description{
Find the mean across Monte Carlo runs.
}
