% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{monte_carlo_mse}
\alias{monte_carlo_mse}
\title{Find the mean square error across Monte Carlo runs.}
\usage{
monte_carlo_mse(per_chain_output, truth)
}
\arguments{
\item{per_chain_output}{Estimate calculated from each chain MCMC in output.}

\item{truth}{The true value being estimated.}
}
\value{
A vector giving the mean square error for each estimate across the chains.
}
\description{
Find the mean square error across Monte Carlo runs.
}
