% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{multiESS_from_output}
\alias{multiESS_from_output}
\title{Calculate multiESS from mcmcse package.}
\usage{
multiESS_from_output(output, chain = 1, rep = 1, external_index = 1)
}
\arguments{
\item{output}{MCMC output.}

\item{chain}{(optional) The chain for which to calculate the multiESS.}

\item{rep}{(optional) The rep (if present) for which to calculate the multiESS.}

\item{external_index}{(optional) The external_index (if present) for which to calculate the multiESS.}
}
\value{
The multiESS for the specified chain/rep.
}
\description{
Calculate multiESS from mcmcse package.
}
