% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{sd_each_chain}
\alias{sd_each_chain}
\title{Find the sd from multiple chains.}
\usage{
sd_each_chain(output, burn_in = 0, chain_index = NULL)
}
\arguments{
\item{burn_in}{(optional) The number of initial iterations to be omitted.}

\item{mcmc_output}{MCMC output, from ilike::load_mcmc_output or otherwise. Can be in tidy format, or in standard nIterations*nVariables format. Both cases must contain a column that labels the chain the output is from.}

\item{chain_column_index}{(optional) The index of the column containing the chain index, needed if there is no column named "chain" or "Chain".}
}
\value{
A list giving the sd of each parameter, for each chain.
}
\description{
Find the sd from multiple chains.
}
