% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_experiments.R
\name{statistic_line_graph}
\alias{statistic_line_graph}
\title{Draw a line graph to show different MCMC statistics across different parameters and methods.}
\usage{
statistic_line_graph(
  statistics,
  x,
  y,
  linetype = NULL,
  colour = NULL,
  log_x = FALSE,
  log_y = FALSE
)
}
\arguments{
\item{statistics}{A data frame containing statistics from across different MCMC experiments.}

\item{x}{The column heading to be used for the x-axis.}

\item{y}{The column heading to be used for the y-axis.}

\item{linetype}{(optional) The column heading to be used for the linetype.}

\item{colour}{(optional) The column heading to be used for the colour.}

\item{log_x}{If TRUE, x-axis values will be logged; if FALSE (default) they won't.}

\item{log_y}{If TRUE, y-axis values will be logged; if FALSE (default) they won't.}
}
\value{
A line graph plot.
}
\description{
Draw a line graph to show different MCMC statistics across different parameters and methods.
}
